Changed token use
Changed the token to be saved in sessionStorage instead of local storage.

Also added id, and permissions to token, so they are no longer available in localStorage as they can be changed.

When you get the code you need to run npm install in order to get
the token decoder.

Made some changes to donation and donor service